# internal/apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
        },
        "description": "API for managing AIG Custom Platform AI security scanning tasks. Based on Tencent Zhuque Lab AI-Infra-Guard (https://github.com/Tencent/AI-Infra-Guard).",
        "title": "AIG Custom Platform 任务API",
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/knowledge/fingerprints/{name}/raw": {
            "get": {
                "description": "Returns exact UTF-8 YAML bytes by one unique recursive info.name match inside the fixed fingerprint root. Ambiguous matches, symlinks, paths outside the root, and files over 1 MiB fail closed.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Legacy envelope whose data contains only content."
                    },
                    "400": {
                        "description": "Invalid opaque name or unsafe file resolution."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Password change is incomplete or role is unsupported."
                    },
                    "404": {
                        "description": "Rule content is absent."
                    },
                    "413": {
                        "description": "Rule content exceeds 1 MiB."
                    },
                    "500": {
                        "description": "Sanitized file-read failure; no path is returned."
                    }
                },
                "summary": "Read exact fingerprint source",
                "tags": [
                    "knowledge"
                ]
            }
        },
...
        "/api/v1/knowledge/vulnerabilities/{id}/raw": {
            "get": {
                "description": "Returns the exact UTF-8 YAML bytes from the fixed Chinese vulnerability root without reserialization. Ambiguous matches, symlinks, paths outside the root, and files over 1 MiB fail closed.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Legacy envelope whose data contains only content."
                    },
                    "400": {
                        "description": "Invalid opaque ID, ambiguous match, or unsafe file resolution."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Password change is incomplete or role is unsupported."
                    },
                    "404": {
                        "description": "Rule content is absent."
                    },
                    "413": {
                        "description": "Rule content exceeds 1 MiB."
                    },
                    "500": {
                        "description": "Sanitized file-read failure; no path is returned."
                    }
                },
                "summary": "Read exact vulnerability source",
                "tags": [
                    "knowledge"
                ]
            }
        },
...
        "/api/v1/knowledge/evaluations/{name}/raw": {
            "get": {
                "description": "Returns the exact UTF-8 JSON bytes through the legacy envelope without reserialization. The requested regular file must stay below 1 MiB and inside the fixed evaluation root; symlinks are rejected.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Legacy envelope whose data contains only content."
                    },
                    "400": {
                        "description": "Invalid opaque name or unsafe file resolution."
                    },
                    "401": {
                        "description": "Unauthenticated."
                    },
                    "403": {
                        "description": "Password change is incomplete or role is unsupported."
                    },
                    "404": {
                        "description": "Evaluation content is absent."
                    },
                    "413": {
                        "description": "Evaluation content exceeds 1 MiB."
                    },
                    "500": {
                        "description": "Sanitized file-read failure; no path is returned."
                    }
                },
                "summary": "Read exact evaluation source",
                "tags": [
                    "knowledge"
                ]
            }
        },
...
    },
...
}